# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: ploykit API
  version: 0.3.0
  description: |
    ploykit 框架 + example 产品应用的 HTTP 契约。

    认证两通道（webx.Authenticate）：
    - 浏览器会话 Cookie `tk_auth`（写请求须带 X-CSRF-Token，/config 拉取 masked token）；
    - PAT `Authorization: Bearer tk_*`（豁免 CSRF）。

    扁平路由（/api/billing/*、/api/usage、/api/audit*、/api/tasks*）的工作区上下文
    由 `X-Workspace-Id` 请求头注入（@ploykit/client 自动携带）。

    错误信封统一为 ErrorBody（webx.ErrorBody）：`{"error":"E_*","message":string,"details"?}`。
security:
  - sessionCookie: []
  - patBearer: []
...
paths:
...
  /api/workspaces/{workspaceId}:
    patch:
      operationId: renameWorkspace
      tags: [workspace]
      summary: 工作区改名（workspace:update）
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameWorkspaceRequest'
      responses:
        '200':
          description: 更新后的工作区
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
    delete:
      operationId: deleteWorkspace
      tags: [workspace]
      summary: 删除工作区（workspace:delete，owner 级动作）
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '204':
          description: 已删除
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'

  # ---------- workspace：成员 / 离开 ----------
  /api/workspaces/{id}/members:
    get:
      operationId: listMembers
      tags: [workspace]
      summary: 成员列表（members:read；分页信封）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
          description: 页码（默认 1）
        - name: page_size
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
          description: 每页条数（默认 50，上限 200）
      responses:
        '200':
          description: 成员列表（分页信封：裸数组 → {items,total}）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPage'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/workspaces/{id}/members/{userId}:
    patch:
      operationId: updateMemberRole
      tags: [workspace]
      summary: 改成员角色（members:write；不能改自己的角色）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRequest'
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
    delete:
      operationId: removeMember
      tags: [workspace]
      summary: 移除成员（members:remove）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: 已移除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOK'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/workspaces/{id}/leave:
    post:
      operationId: leaveWorkspace
      tags: [workspace]
      summary: 退出工作区（last-owner 拒绝）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
      responses:
        '200':
          description: 已退出
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOK'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/DefaultError'
  # tenant 侧所有权转让，
  # 与 POST /api/admin/workspaces/{id}/transfer-ownership 同域方法双入口。
  /api/workspaces/{id}/transfer-ownership:
    post:
      operationId: transferOwnership
      tags: [workspace]
      summary: 转让所有权（owner 发起；单事务目标升 owner、原 owner 降 member）
      description: |
        发起者须为本工作区 owner（守卫在 workspace 域，不经 authz 权限矩阵——
        所有权不是成员管理权）。GitLab"至少一个 human owner"语义在域层落地：
        先升后降单事务，任何中途/终态保有 ≥1 owner。域钩子 BeforeOwnerChange
        （Validating 可阻断）/OnOwnerTransfer（Observational）与审计
        workspace.owner_transfer 照旧。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, description: 新 owner（须是本工作区既有活跃成员且账号未禁用） }
      responses:
        '200':
          description: 已转让
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
          description: 发起者非本工作区 owner
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
          description: 目标已是 owner/转给自己/目标账号 disabled/并发竞态/BeforeOwnerChange 拦截
        default:
          $ref: '#/components/responses/DefaultError'

  # ---------- workspace：邀请 ----------
  /api/workspaces/{id}/invitations:
    get:
      operationId: listInvitations
      tags: [workspace]
      summary: 邀请列表（invites:read；仅 pending 且未过期；分页信封）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
          description: 页码（默认 1）
        - name: page_size
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
          description: 每页条数（默认 50，上限 200）
      responses:
        '200':
          description: 邀请列表（分页信封：裸数组 → {items,total}）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationPage'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
    post:
      operationId: createInvitation
      tags: [workspace]
      summary: 发邀请（members:invite）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        '201':
          description: 已创建（邮件由服务端异步发送）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/workspaces/{id}/invitations/{invId}:
    delete:
      operationId: revokeInvitation
      tags: [workspace]
      summary: 撤回邀请（members:invite）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
        - name: invId
          in: path
          required: true
          schema: { type: string }
          description: 邀请 ID
      responses:
        '204':
          description: 已撤回
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/invitations/mine:
    get:
      operationId: myInvitations
      tags: [workspace]
      summary: 收到的待处理邀请（带工作区名）
      responses:
        '200':
          description: 邀请列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitationWithWorkspace'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/invitations/{id}/accept:
    post:
      operationId: acceptInvitation
      tags: [workspace]
      summary: 接受邀请（加入工作区）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 邀请 ID
      responses:
        '200':
          description: 加入的工作区
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/invitations/{id}/decline:
    post:
      operationId: declineInvitation
      tags: [workspace]
      summary: 拒绝邀请
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 邀请 ID
      responses:
        '200':
          description: 已拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOK'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'

  # ---------- workspace：分享链接 ----------
  /api/workspaces/{id}/share-links:
    get:
      operationId: listShareLinks
      tags: [workspace]
      summary: 分享链接列表（invites:read；仅未撤销且未过期；分页信封）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
          description: 页码（默认 1）
        - name: page_size
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
          description: 每页条数（默认 50，上限 200）
      responses:
        '200':
          description: 分享链接列表（分页信封：裸数组 → {items,total}）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLinkPage'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
    post:
      operationId: createShareLink
      tags: [workspace]
      summary: 创建分享链接（members:invite；完整 code 仅本次响应出现一次）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareLinkRequest'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateShareLinkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/workspaces/{id}/share-links/{linkId}:
    delete:
      operationId: revokeShareLink
      tags: [workspace]
      summary: 作废分享链接（members:invite）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作区 ID
        - name: linkId
          in: path
          required: true
          schema: { type: string }
          description: 分享链接 ID
      responses:
        '204':
          description: 已作废
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/share-links/redeem:
    post:
      operationId: redeemShareLink
      tags: [workspace]
      summary: 兑换分享链接码加入工作区
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeemShareLinkRequest'
      responses:
        '200':
          description: 加入的工作区
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/DefaultError'

  # ---------- billing ----------
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tk_auth
      description: 浏览器会话 Cookie（webx.Authenticate 签发/续期；写请求须另带 X-CSRF-Token）
    patBearer:
      type: http
      scheme: bearer
      description: '个人访问令牌（Authorization: Bearer tk_*，CSRF 豁免）'
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string }
      description: 工作区 ID
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
      description: 成员用户 ID
...
  responses:
    DefaultError:
      description: 其他错误（统一 ErrorBody 信封）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    BadRequest:
      description: 参数非法（E_VALIDATION / E_BAD_JSON / E_PAYLOAD_TOO_LARGE）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthenticated:
      description: 未认证（E_UNAUTHENTICATED）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Forbidden:
      description: 无权限（E_FORBIDDEN）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    NotFound:
      description: 不存在（E_NOT_FOUND）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Conflict:
      description: 冲突（E_CONFLICT）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
  schemas:
...
    ErrorBody:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: 错误码（E_UNAUTHENTICATED / E_FORBIDDEN / E_NOT_FOUND / E_VALIDATION / E_CONFLICT / E_RATE_LIMITED / E_QUOTA_EXCEEDED / E_INTERNAL / E_TIMEOUT / E_PAYLOAD_TOO_LARGE / E_UNAVAILABLE）
        message:
          type: string
        details:
          description: 可选补充（通常为字符串详情）
    StatusOK:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
...
    Workspace:
      type: object
      required: [id, slug, name, plan_code, created_at]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        plan_code:
          type: string
        created_at:
          type: string
          format: date-time
...
    Member:
      type: object
      required: [user_id, email, display_name, role, created_at]
      properties:
        user_id:
          type: string
        email:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
        created_at:
          type: string
          format: date-time
    Invitation:
      type: object
      required: [id, workspace_id, email, role, status, expires_at, created_at]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
        status:
          type: string
          enum: [pending, accepted, declined, revoked, expired]
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    InvitationWithWorkspace:
      allOf:
        - $ref: '#/components/schemas/Invitation'
        - type: object
          required: [workspace_name, workspace_slug]
          properties:
            workspace_name:
              type: string
            workspace_slug:
              type: string
    ShareLink:
      type: object
      required: [id, code_prefix, role, max_uses, uses, expires_at, created_at]
      properties:
        id:
          type: string
        code_prefix:
          type: string
          description: 展示用前缀（完整 code 仅创建响应出现一次）
        role:
          type: string
          enum: [admin, member]
        max_uses:
          type: integer
        uses:
          type: integer
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    MemberPage:
      type: object
      description: 成员列表分页信封（total 为过滤命中总数，非本页条数）
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        total:
          type: integer
    InvitationPage:
      type: object
      description: 邀请列表分页信封（total 为过滤命中总数，非本页条数）
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Invitation'
        total:
          type: integer
    ShareLinkPage:
      type: object
      description: 分享链接列表分页信封（total 为过滤命中总数，非本页条数）
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ShareLink'
        total:
          type: integer
...
    RenameWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateMemberRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, member]
    CreateInvitationRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, member]
    CreateShareLinkRequest:
      type: object
      required: [role, max_uses, ttl_hours]
      properties:
        role:
          type: string
          enum: [admin, member]
        max_uses:
          type: integer
        ttl_hours:
          type: integer
    CreateShareLinkResponse:
      type: object
      required: [link, code]
      properties:
        link:
          $ref: '#/components/schemas/ShareLink'
        code:
          type: string
          description: 完整兑换码（仅本次出现一次）
    RedeemShareLinkRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string

    # ---- billing（PKPlan/PKOrder/Subscription 对齐 packages/ui） ----
...